Fix autoloading for charts after history swap
