+++ resources/js/app.js
@@ -10,6 +10,7 @@ import 'tinymce/themes/silver/theme.min.js';
 import 'tinymce/models/dom/model.min.js';
 import './components/autocomplete-input';
 import './components/taxonomy-create';
+import './components/chart';
 import { Chart, registerables } from 'chart.js';
 
 Chart.register(...registerables);
